% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eastweb_data.R
\name{eastweb_data}
\alias{eastweb_data}
\title{Get data  from an EASTWeb database}
\usage{
eastweb_data(project_name, plugin = c("TRMM3B42", "TRMM3b42RT", "ModisLST",
  "ModisNBAR", "NldasNOAH", "NldasForcing"), user, password,
  db_name = "epidemiadb", host = "localhost", port = 5432)
}
\arguments{
\item{project_name}{Name of the project. Does not need underscores.}

\item{plugin}{Name of the plugin. Must be one of "TRMM3B42", "TRMM3b42RT",
"ModisLST", "ModisNBAR", "NldasNOAH", or "NldasForcing".}

\item{user}{Username.}

\item{password}{Password.}

\item{db_name}{Name of the database.}

\item{host}{Host name. Defaults to "localhost".}

\item{port}{Port number. Defaults to the postgreSQL default of 5432.}
}
\value{
A data frame containing the zonal statistics table for the given
plugin in an EASTWeb project. Even if a project has multiple plugins, each
one must be queried seperately.
}
\description{
Get data  from an EASTWeb database
}
\examples{

}

