% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eastweb_projects.R
\name{eastweb_global_tables}
\alias{eastweb_global_tables}
\title{Get global schema tables in an EASTWeb database}
\usage{
eastweb_global_tables(db_name, user, password, host = "localhost",
  port = 5432)
}
\arguments{
\item{db_name}{Name of the database.}

\item{user}{Username}

\item{password}{Password}

\item{host}{Host name. Defaults to "localhost".}

\item{port}{Port number. Defaults to the postgreSQL default of 5432}
}
\value{
Returns a list of data frames, one for each table in the specified
database.
}
\description{
Get global schema tables in an EASTWeb database
}
\examples{

}

