% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_names.R
\name{format_project_name}
\alias{format_project_name}
\title{Format a project name as it appears in the Project table in an EASTWeb
database}
\usage{
format_project_name(project_name)
}
\arguments{
\item{project_name}{The name of the project. Case sensitive.}
}
\value{
The formatted named of the project. Spaces are converted to
underscores.
}
\description{
Format a project name as it appears in the Project table in an EASTWeb
database
}
\examples{
eastweb_proj_name("Project Name")

}

