% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_names.R
\name{format_schema_name}
\alias{format_schema_name}
\title{Format a schema name as it appears in the an EASTWeb
database}
\usage{
format_schema_name(project_name, plugin = c("TRMM3B42", "TRMM3b42RT",
  "ModisLST", "ModisNBAR", "NldasNOAH", "NldasForcing"))
}
\arguments{
\item{project_name}{The name of the project. Case sensitive.}

\item{plugin}{The name of the plugin. Case sensitive.}
}
\value{
The formatted named of the schema. The project name is pasted
together with the plugin name and everything is converted to lower case.
}
\description{
Format a schema name as it appears in the an EASTWeb
database
}
\examples{
format_schema_name("Project Name", "PluginName")

}

