% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_schema.R
\name{read_global_schema}
\alias{read_global_schema}
\title{Read global schema using an open connection to an EASTWeb database}
\usage{
read_global_schema(db_con)
}
\arguments{
\item{db_con}{A connection to an EASTWeb database}
}
\value{
Returns a list of data frames, one for each table in the specified
database.
}
\description{
Read global schema using an open connection to an EASTWeb database
}
\examples{

}

