% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_summary.R
\name{read_project_summary}
\alias{read_project_summary}
\title{Read information for a particular project using an open connection to an
EASTWeb database}
\usage{
read_project_summary(db_con, project_name)
}
\arguments{
\item{db_con}{A connection to an EASTWeb database}

\item{project_name}{Name of the project to extract.}
}
\value{
A list with values describing the project.
}
\description{
Read information for a particular project using an open connection to an
EASTWeb database
}
\examples{

}

