% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{read_projects}
\alias{read_projects}
\title{Read all project information using an open connection to an EASTWeb database}
\usage{
read_projects(db_con)
}
\arguments{
\item{db_con}{A connection to an EASTWeb database}
}
\value{
Returns a dataframe of the projects in the specified database.
}
\description{
Read all project information using an open connection to an EASTWeb database
}
\examples{

}

